+++ src/demo/xmake.lua
@@ -93,3 +93,4 @@ target("demo")
     if is_plat("bsd") then
         add_syslinks("execinfo")
     end
+